Refactored syscall mod into crate
