+++ src/business_logic/transaction/error.rs
@@ -1,12 +1,10 @@
 use crate::{
     business_logic::execution::os_usage::OsResources,
-    core::errors::{
-        contract_address_errors::ContractAddressError, state_errors::StateError,
-        syscall_handler_errors::SyscallHandlerError,
-    },
+    core::errors::{contract_address_errors::ContractAddressError, state_errors::StateError},
     definitions::transaction_type::TransactionType,
     starknet_runner::starknet_runner_error::StarknetRunnerError,
     starkware_utils::starkware_errors::StarkwareError,
+    syscalls::syscall_handler_errors::SyscallHandlerError,
     utils::ClassHash,
 };
 use cairo_vm::{
