+++ src/business_logic/transaction/objects/internal_deploy.rs
@@ -10,7 +10,6 @@ use crate::{
     },
     core::{
         contract_address::starknet_contract_address::compute_deprecated_class_hash,
-        errors::syscall_handler_errors::SyscallHandlerError,
         transaction_hash::starknet_transaction_hash::calculate_deploy_transaction_hash,
     },
     definitions::{
@@ -20,6 +19,7 @@ use crate::{
     hash_utils::calculate_contract_address,
     services::api::contract_classes::deprecated_contract_class::{ContractClass, EntryPointType},
     starkware_utils::starkware_errors::StarkwareError,
+    syscalls::syscall_handler_errors::SyscallHandlerError,
     utils::{calculate_tx_resources, felt_to_hash, Address, ClassHash},
 };
 use cairo_vm::felt::Felt252;
