+++ src/core/errors/mod.rs
@@ -1,3 +1,2 @@
 pub mod contract_address_errors;
 pub mod state_errors;
-pub mod syscall_handler_errors;
