+++ src/core/mod.rs
@@ -1,5 +1,4 @@
 pub mod block_hash;
 pub mod contract_address;
 pub mod errors;
-pub mod syscalls;
 pub mod transaction_hash;
