+++ src/core/transaction_hash/starknet_transaction_hash.rs
@@ -2,11 +2,11 @@ use crate::{
     core::{
         contract_address::starknet_contract_address::compute_deprecated_class_hash,
         contract_address::v2::starknet_sierra_contract_address::compute_sierra_class_hash,
-        errors::syscall_handler_errors::SyscallHandlerError,
     },
     definitions::constants::CONSTRUCTOR_ENTRY_POINT_SELECTOR,
     hash_utils::compute_hash_on_elements,
     services::api::contract_classes::deprecated_contract_class::ContractClass,
+    syscalls::syscall_handler_errors::SyscallHandlerError,
     utils::Address,
 };
 use cairo_lang_starknet::contract_class::ContractClass as SierraContractClass;
