+++ src/parser_errors.rs
@@ -1,10 +1,8 @@
 use crate::{
     business_logic::transaction::error::TransactionError,
-    core::errors::{
-        contract_address_errors::ContractAddressError, state_errors::StateError,
-        syscall_handler_errors::SyscallHandlerError,
-    },
+    core::errors::{contract_address_errors::ContractAddressError, state_errors::StateError},
     services::api::contract_classes::deprecated_contract_class::EntryPointType,
+    syscalls::syscall_handler_errors::SyscallHandlerError,
 };
 use thiserror::Error;
