+++ src/starknet_runner/runner.rs
@@ -1,6 +1,6 @@
 use super::starknet_runner_error::StarknetRunnerError;
 use crate::business_logic::transaction::error::TransactionError;
-use crate::core::syscalls::syscall_handler::HintProcessorPostRun;
+use crate::syscalls::syscall_handler::HintProcessorPostRun;
 use cairo_lang_starknet::casm_contract_class::CasmContractClass;
 use cairo_vm::felt::Felt252;
 use cairo_vm::hint_processor::hint_processor_definition::HintProcessor;
@@ -386,7 +386,7 @@ mod test {
             fact_state::in_memory_state_reader::InMemoryStateReader,
             state::cached_state::CachedState, transaction::error::TransactionError,
         },
-        core::syscalls::{
+        syscalls::{
             deprecated_business_logic_syscall_handler::DeprecatedBLSyscallHandler,
             deprecated_syscall_handler::DeprecatedSyscallHintProcessor,
             syscall_handler::SyscallHintProcessor,
