+++ src/syscalls/business_logic_syscall_handler.rs
@@ -3,6 +3,7 @@
 use std::collections::HashMap;
 use std::ops::Add;
 
+use super::syscall_handler_errors::SyscallHandlerError;
 use super::syscall_request::{
     EmitEventRequest, FromPtr, GetBlockHashRequest, GetBlockTimestampRequest, StorageReadRequest,
     StorageWriteRequest,
@@ -33,7 +34,7 @@ use crate::{
             state_api::{State, StateReader},
         },
     },
-    core::errors::{state_errors::StateError, syscall_handler_errors::SyscallHandlerError},
+    core::errors::state_errors::StateError,
     definitions::{
         constants::CONSTRUCTOR_ENTRY_POINT_SELECTOR, general_config::StarknetGeneralConfig,
     },
