+++ src/syscalls/deprecated_business_logic_syscall_handler.rs
@@ -7,6 +7,7 @@ use super::{
         DeprecatedGetTxInfoResponse, DeprecatedGetTxSignatureResponse,
         DeprecatedStorageReadResponse, DeprecatedWriteSyscallResponse,
     },
+    syscall_handler_errors::SyscallHandlerError,
     syscall_info::get_deprecated_syscall_size_from_name,
 };
 use crate::{
@@ -20,7 +21,7 @@ use crate::{
         },
         transaction::error::TransactionError,
     },
-    core::errors::{state_errors::StateError, syscall_handler_errors::SyscallHandlerError},
+    core::errors::state_errors::StateError,
     definitions::general_config::StarknetGeneralConfig,
     hash_utils::calculate_contract_address,
     public::abi::CONSTRUCTOR_ENTRY_POINT_SELECTOR,
@@ -845,7 +846,7 @@ mod tests {
             fact_state::in_memory_state_reader::InMemoryStateReader,
             state::cached_state::CachedState,
         },
-        core::errors::syscall_handler_errors::SyscallHandlerError,
+        syscalls::syscall_handler_errors::SyscallHandlerError,
         utils::{test_utils::*, Address},
     };
     use cairo_vm::felt::Felt252;
