+++ src/syscalls/mod.rs
@@ -6,6 +6,7 @@ pub mod deprecated_syscall_response;
 pub mod hint_code;
 pub mod other_syscalls;
 pub mod syscall_handler;
+pub mod syscall_handler_errors;
 pub mod syscall_info;
 pub mod syscall_request;
 pub mod syscall_response;
